requirements.txt
django
djangorestframework
psycopg2==2.7.7
django-environ==0.4.4
gevent
googlemaps
